don't try to use orderly-web command before it is installed

+++ scripts/stop-orderly-web.sh
@@ -2,4 +2,6 @@
 set -ex
 HERE=$(dirname $0)
 
-orderly-web stop $HERE --kill --force
+if which -a orderly-web >/dev/null; then
+  orderly-web stop $HERE --kill --force
+fi
